+++ packages/botonic-react/src/util/strings.ts
@@ -1,0 +1,6 @@
+export const truncateText = (text, maxLength, ellipsis = '...') => {
+  if (text.length > maxLength) {
+    return text.substring(0, maxLength - ellipsis.length) + ellipsis
+  }
+  return text
+}